feat: add template Main
